Add a resource list component that can list all the user resources
This adds the resource list component as specified in the linked issue. This list will render the resources card instead of simple link like the data version of this component.

Both the no-resources state and the loading state are handled through slots to ensure the maximum level of customization.

Work is related to manifoldco/engineering#8644