Remove the input from `manifold-data-provision-button`
That input can now be added outside the component itself for an easier control of it's position, associated label and stylings. Instead, a `resource-name` attribute must be provided to the provision button for resources to be properly provisioned.

Also added a lot of spec tests to the provision button.

Work is related to manifoldco/engineering#8687