Add two new components for displaying a resource's details
The product and the plan component will display the resource's product and plan respectively thanks to the `resource-container` component. These can be used by platforms to build a resource details page with their own styling thanks to the state tunnel in `resource-container`.

For testing purposes, I also added a `mock-resource` component that allows us or platforms to fake a `resource-container` for testing purposes.

Work is related to manifoldco/engineering#8689