Add a new rename button for resources
This button can be used as a data component standalone to rename a resource. It will require a resource name and can fetch the resource id if not provided. The button expects the new name for the resource for renaming.

It can also be used as part of the `resource-container` component by using its `manifold-resource-rename` version. Both version are stateless and will need a corresponding input to work.

Work is related to manifoldco/engineering#8690