Refactor resource name to resource label for clarity
Everywhere in our docs we are using resource label for the slug we can use to fetch and resource name for the display name. In UI, resource name was the slug and the display name. That distinction has now been fixed and we should have a more clear standard when it comes to names and labels.

This PR also adds a new documentation page for how to create a fully connected resource details view with the components I have been creating.

Work is related to manifoldco/engineering#8729