Add the SSO button and its code to UI
The SSO button is a data component that will trigger a POST sso request on connector and return the redirect_url necessary for properly logging into the resource's product dashboard.

It also adds a `manifold-resource-sso` wrapper components that will automatically inject some of the attributes into the button.

Work is related to manifoldco/engineering#8691

Test with storybook