Add a wait of 15 seconds for the token to exist during fetches
When a request for fetching data on graphql or rest is sent to the servers, all the async functions will now wait for a maximum of 15 seconds (Configurable) for the token to exist before returning an error.

All the components using rest have been modified to now use the function added to the manifold-connection. Tests have been adapted to make sure the rest and graphql functions are properly injected in components.

Work is related to manifoldco/engineering#8858