Rework how the resources list work with operations
We now use the code from the ZEIT hackahton, slightly modified to fit our needs. This is still FAR from perfect, but the list will now properly show provisioning and resizing operations if they exist. It will also show the data we want when deprovisoning.

The optimal next step right now would be to implement a GraphQl endpoint and rework components that use the resources rest endpoints to use GraphQl and not worry about operations anymore.

Work is related to manifoldco/engineering#8832

Try testing with the render dashboard, it should show the proper behavior.