Add a timeout on the resource container to keep fetching until valid
This will enable the container to fetch a resource that is still provisioning or is having an operation run on it. It will keep fetching until "done" to ensure that the container can handle cases like being rendered on a provisioning resource.

Work is related to manifoldco/engineering#8832

Test with the render dashboard after provisioning a resource to see it in action.