Fix the service card title dynamic size
The dynamic sizing the the service card's first row made it so that the description would often look misaligned when products with different a number of lines for the description were put next to each other. This adds a fixed width for the first row of the service card.

Work is related to manifoldco/engineering#8934