Update the auth-token event name and document how it works
This changes the `manifold-auth-token` event name from whatever stencil chose to a fixed `manifold-token-received`. This will also ensure that this event can bubble up to the document rather than stay locked to the component.

This PR also updates the docs to better document the auth token component with the new shadowcat behavior.