Expose the credentials buttons colors for more customization
Work is related to manifoldco/engineering#8940

This gives platforms the ability to customize the credentials buttons to their theme without needing too many CSS variables. The prop is passed through all the wrapper components to the credentials view component.

Test with the docs and set the two properties manually in the dom.