Make the credentials use slots and have a default value with the buttons
This makes use of slots and query selector to allow the behavior where, if you provide a slot's content, it will render that slot, otherwise it will render the default object.

Due to the forward slot component, we couldn't use the browser native ability to do that without any element checking.