Fix the service card loading to keep loading when fetching if free (#355)
* Fix the service card loading to keep loading when fetching if free

The service card would render then start loading if the product is free, which led to the free badge jumping around a lot. Products will now load in different intervals, but will always be loaded with their badges on the first render.

Work is related to manifoldco/engineering#8935

* Fix tests and make sure that the marketplace prevents free fetching

The fetching will now only happen if, and only if, the is-free property is not already set on the service card. This prevents some double fetching that was happening previously.

* Add changelog

* Remove a console.log and fix a happo test

* Unfix happo

* Use manifold-service-card-view in manifold-marketplace

* Fix Happo tests

* Fix Lint

* Fix accidental commits from master rebase

* Add manifold-resource-product Happo test

* Add happo tests