Expose the credentials buttons colors for more customization (#362)
* Expose the credentials buttons colors for more customization

Work is related to manifoldco/engineering#8940

This gives platforms the ability to customize the credentials buttons to their theme without needing too many CSS variables. The prop is passed through all the wrapper components to the credentials view component.

Test with the docs and set the two properties manually in the dom.

* Fix small mistakes and internal component name

* Remove color change in docs

* Fix happo tests

* Add happo tests for the custom button colors

* Make the credentials use slots and have a default value with the buttons

This makes use of slots and query selector to allow the behavior where, if you provide a slot's content, it will render that slot, otherwise it will render the default object.

Due to the forward slot component, we couldn't use the browser native ability to do that without any element checking.