Mock resource endpoints in the docs
This PR hardcodes some data into the docs and mock all fetch calls made to the LIST resources endpoint and the LIST operations endpoint. The goal is to make the list resource page work properly without having to implement an auth system in our docs.

Test with the docs, view the list resources component's page.