Added a new get credentials data button
Work is related to manifoldco/engineering#9042

This button will fetch the credentials of a resource on click and then return those credentials inside of a custom event triggered on success. If any error happened during the fetch process, an error will be triggered instead

Test with storybook or the docs, if you are authenticated, the button should fetch the credentials.