Convert the manifold-resource-plan component to GraphQL
This converts the manifold-resource-plan component to GraphQL in an effort to convert the manifold-resource-container component. It does not actually uses the graphQL data, but instead uses the plan data conversion function created by dave.js.

Work is related to manifoldco/engineering#9192

Test with storybook, I suggest using the `manifold-mock-resource` component in the resource stories for easy testing.