An action to run TestCafe on StoryBook deployment
This is a first pass at a GitHub action for running TestCafe against the StoryBook previews that appear in pull requests. Currently, this action will only target sslotsky/test-cafe.

As soon as we achieve a successful run, we can merge sslotsky/test-cafe and update the action to run against pull request branches, or even the master branch.